Filter: use only single lower bit of filter return value
Implementation change: use a uint64_t as filter return value, and test
with a mask for this bit. This reserves other bits for future use.

+++ liblttng-ust/lttng-filter-interpreter.c
@@ -105,7 +105,7 @@ int stack_strcmp(struct estack *stack, int top, const char *cmp_type)
 	return diff;
 }
 
-int lttng_filter_false(void *filter_data,
+uint64_t lttng_filter_false(void *filter_data,
 		const char *filter_stack_data)
 {
 	return 0;
@@ -157,7 +157,12 @@ LABEL_##name
 
 #endif
 
-int lttng_filter_interpret_bytecode(void *filter_data,
+/*
+ * Return 0 (discard), or raise the 0x1 flag (log event).
+ * Currently, other flags are kept for future extensions and have no
+ * effect.
+ */
+uint64_t lttng_filter_interpret_bytecode(void *filter_data,
 		const char *filter_stack_data)
 {
 	struct bytecode_runtime *bytecode = filter_data;
@@ -279,6 +284,7 @@ int lttng_filter_interpret_bytecode(void *filter_data,
 			goto end;
 
 		OP(FILTER_OP_RETURN):
+			/* LTTNG_FILTER_DISCARD  or LTTNG_FILTER_RECORD_FLAG */
 			retval = !!estack_ax_v;
 			ret = 0;
 			goto end;
